requirements.txt
flask==3.0.2
flask-socketio==5.3.6
openai>=2.5.0
python-dotenv==1.0.1
eventlet==0.36.1 
urllib3==2.5.0
bs4==0.0.2
rapidfuzz==3.14.1
scikit-learn==1.7.2
httpx>=0.28.1
httpcore==1.0.9
h11==0.16.0